+++ tests/Ardalis.Specification.IntegrationTests/SampleSpecs/PostsByBlogPaginatedSpec.cs
@@ -10,19 +10,5 @@ namespace Ardalis.Specification.IntegrationTests.SampleSpecs
             ApplyPaging(skip, take);
         }
     }
-    public class PostsByBlogOrderedSpec : BaseSpecification<Post>
-    {
-        public PostsByBlogOrderedSpec(int blogId, bool isAscending = true)
-            : base(p => p.BlogId == blogId)
-        {
-            if(isAscending)
-            {
-                ApplyOrderBy(p => p.Id);
-            }
-            else
-            {
-                ApplyOrderByDescending(p => p.Id);
-            }
-        }
-    }
+
 }
